float

+++ js/script.js
@@ -17,9 +17,9 @@ if (navigator.serviceWorker) {
 
 function enterClicked() {
   // input
-  const length = parseInt(document.getElementById("length-of-pyramid").value)
-  const width = parseInt(document.getElementById("width-of-Pyramid").value)
-  const height = parseInt(document.getElementById("height-of-pyramid").value)
+  const length = parseFloat(document.getElementById("length-of-pyramid").value)
+  const width = parseFloat(document.getElementById("width-of-Pyramid").value)
+  const height = parseFloat(document.getElementById("height-of-pyramid").value)
   // process
   const volume = (length * width * height) / 3
